+++ web-wallet/CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix Receive tab content overflows [#1901]
 - Add missing "Soehne Mono" and its `@font-face` definition [#2071]
 - The sync promise should be set to `null` after aborting a sync [#2118]
+- Fix rounding errors in migration amount input [#2303]
 
 ## [0.5.0] - 2024-03-27
 
@@ -251,6 +252,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#2118]: https://github.com/dusk-network/rusk/issues/2175
 [#2196]: https://github.com/dusk-network/rusk/issues/2196
 [#2014]: https://github.com/dusk-network/rusk/issues/2014
+[#2303]: https://github.com/dusk-network/rusk/issues/2303
 
 <!-- VERSIONS -->